Adding visible sha256 and tests
