+++ Tests/CryptoTests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="nanoFramework.CoreLibrary" version="1.12.0-preview.9" targetFramework="netnanoframework10" />
+  <package id="nanoFramework.TestFramework" version="1.0.178" targetFramework="netnanoframework10" developmentDependency="true" />
+</packages>
