+++ nanoFramework.Azure.Devices.Client.nuspec
@@ -30,7 +30,6 @@ This is an SDK for Azure IoT Hub using MQTT broker.</description>
       <dependency id="nanoFramework.System.Net.Http" version="1.4.0-preview.26" />
       <dependency id="nanoFramework.System.Text" version="1.1.3-preview.15" />
       <dependency id="nanoFramework.System.Threading" version="1.0.4-preview.16" />
-      <dependency id="nanoFramework.Windows.Storage.Streams" version="1.12.4-preview.16" />
     </dependencies>
   </metadata>
   <files>
