+++ nanoFramework.Azure.Devices.Client/packages.config
@@ -11,6 +11,5 @@
   <package id="nanoFramework.System.Net.Http" version="1.4.0-preview.26" targetFramework="netnanoframework10" />
   <package id="nanoFramework.System.Text" version="1.1.3-preview.15" targetFramework="netnanoframework10" />
   <package id="nanoFramework.System.Threading" version="1.0.4-preview.16" targetFramework="netnanoframework10" />
-  <package id="nanoFramework.Windows.Storage.Streams" version="1.12.4-preview.16" targetFramework="netnanoframework10" />
   <package id="Nerdbank.GitVersioning" version="3.4.220" developmentDependency="true" targetFramework="netnanoframework10" />
 </packages>
